[UPD] ore: continue refactoring
