+++ demo_ore/__manifest__.py
@@ -7,12 +7,13 @@
     "depends": ["ore", "ore_data", "muk_branding"],
     "data": [
         "data/res_partner.xml",
+        "data/ore_membre.xml",
         "data/res_company.xml",
         "data/user_demo.xml",
         "data/ore_offre_service.xml",
         "data/ore_demande_service.xml",
         "data/ore_echange_service.xml",
-        "data/res_partner_favoris.xml",
+        "data/ore_membre_favoris.xml",
     ],
     "installable": True,
     "post_init_hook": "post_init_hook",
