+++ demo_ore/data/ore_membre_favoris.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <data noupdate="1">
+        <record id="ore_membre_favoris_administrateur_mathieu_benoit" model="ore.membre.favoris">
+            <field name="membre_id" ref="demo_ore.base_partner_admin" />
+        </record>
+        <record id="ore_membre_favoris_martin_petit" model="ore.membre.favoris">
+            <field name="membre_id" ref="ore_membre_demo_martin_petit" />
+        </record>
+        <record id="ore_membre_favoris_denis_lemarchand" model="ore.membre.favoris">
+            <field name="membre_id" ref="ore_membre_demo_martin_bergeron" />
+        </record>
+        <record id="ore_membre_favoris_alice_poitier" model="ore.membre.favoris">
+            <field name="membre_id" ref="ore_membre_demo_alice_poitier" />
+        </record>
+    </data>
+</odoo>
