--- demo_ore/data/res_partner_favoris.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <data noupdate="1">
-        <record id="res_partner_favoris_administrateur_mathieu_benoit" model="res.partner.favoris">
-            <field name="membre_id" ref="demo_ore.base_partner_admin" />
-        </record>
-        <record id="res_partner_favoris_martin_petit" model="res.partner.favoris">
-            <field name="membre_id" ref="partner_demo_martin_petit" />
-        </record>
-        <record id="res_partner_favoris_denis_lemarchand" model="res.partner.favoris">
-            <field name="membre_id" ref="partner_demo_denis_lemarchand" />
-        </record>
-        <record id="res_partner_favoris_alice_poitier" model="res.partner.favoris">
-            <field name="membre_id" ref="partner_demo_alice_poitier" />
-        </record>
-    </data>
-</odoo>
