+++ demo_website_chat_ore/data/ore_chat_group.xml
@@ -2,7 +2,7 @@
 <odoo>
     <data noupdate="1">
         <record id="msg_group_1" model="ore.chat.group">
-            <field name="membre_ids" eval="[(4, ref('demo_ore.base_partner_admin')), (4, ref('demo_ore.partner_demo_alice_poitier'))]" />
+            <field name="membre_ids" eval="[(4, ref('demo_ore.base_partner_admin')), (4, ref('demo_ore.ore_membre_demo_alice_poitier'))]" />
         </record>
     </data>
 </odoo>
